+++ modules/environment_modeling/libraries/libWPG/testing/google_test/source_code/Test_WorldPointGeneratorCuboid.cpp
@@ -68,7 +68,7 @@ TEST_NUMBEROFWORLDPOINTS_DEFAULTCONSTRUCTOR_NOGENERATION_ISZERO(WorldPointGenera
 ///////////////////////////////////////////////////////////////////////////////
 TEST_NUMBEROFWORLDPOINTS_DEFAULTCONSTRUCTOR_ISMATCHING(WorldPointGeneratorCuboid, Test_NumberOfWorldPoints_DefaultConstructor_IsMatching)
 {
-    const uint64 NumberOfWorldPointsToCreate{1000U};
+    const uint32 NumberOfWorldPointsToCreate{1000U};
 
     WorldPointGeneratorCuboid WPG;
     WPG.GeneratePointCloud();
@@ -86,7 +86,7 @@ TEST_NUMBEROFWORLDPOINTS_DEFAULTCONSTRUCTOR_ISMATCHING(WorldPointGeneratorCuboid
 ///////////////////////////////////////////////////////////////////////////////
 TEST_NUMBEROFWORLDPOINTS_50_NOGENERATION_ISZERO(WorldPointGeneratorCuboid, Test_NumberOfWorldPoints_50_NoGeneration_IsZero)
 {
-    const uint64 NumberOfWorldPointsToCreate{50};
+    const uint32 NumberOfWorldPointsToCreate{50};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate);
 
@@ -103,7 +103,7 @@ TEST_NUMBEROFWORLDPOINTS_50_NOGENERATION_ISZERO(WorldPointGeneratorCuboid, Test_
 ///////////////////////////////////////////////////////////////////////////////
 TEST_NUMBEROFWORLDPOINTS_100_NOGENERATION_ISZERO(WorldPointGeneratorCuboid, Test_NumberOfWorldPoints_100_NoGeneration_IsZero)
 {
-    const uint64 NumberOfWorldPointsToCreate{100};
+    const uint32 NumberOfWorldPointsToCreate{100};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate);
 
@@ -118,7 +118,7 @@ TEST_NUMBEROFWORLDPOINTS_100_NOGENERATION_ISZERO(WorldPointGeneratorCuboid, Test
 ///////////////////////////////////////////////////////////////////////////////
 TEST_NUMBEROFWORLDPOINTS_50_ISMATCHING(WorldPointGeneratorCuboid, Test_NumberOfWorldPoints_50_IsMatching)
 {
-    const uint64 NumberOfWorldPointsToCreate{50U};
+    const uint32 NumberOfWorldPointsToCreate{50U};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate);
     WPG.GeneratePointCloud();
@@ -134,7 +134,7 @@ TEST_NUMBEROFWORLDPOINTS_50_ISMATCHING(WorldPointGeneratorCuboid, Test_NumberOfW
 ///////////////////////////////////////////////////////////////////////////////
 TEST_NUMBEROFWORLDPOINTS_100_ISMATCHING(WorldPointGeneratorCuboid, Test_NumberOfWorldPoints_100_IsMatching)
 {
-    const uint64 NumberOfWorldPointsToCreate{100U};
+    const uint32 NumberOfWorldPointsToCreate{100U};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate);
     WPG.GeneratePointCloud();
@@ -241,14 +241,14 @@ TEST_WORLDPOINT_4_DEFAULTCONSTRUCTOR_ISMATCHING(WorldPointGeneratorCuboid, Test_
 ///////////////////////////////////////////////////////////////////////////////
 TEST_WORLDPOINT_0_SEED_10_ISMATCHING(WorldPointGeneratorCuboid, Test_WorldPoint_0_Seed_10_IsMatching)
 {
-    const uint64  NumberOfWorldPointsToCreate{1000U};
+    const uint32  NumberOfWorldPointsToCreate{1000U};
     const float64 MinX{-5.0};
     const float64 MaxX{5.0};
     const float64 MinY{-2.0};
     const float64 MaxY{2.0};
     const float64 MinZ{3.0};
     const float64 MaxZ{30.0};
-    const uint64  Seed{10U};
+    const uint32  Seed{10U};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate, MinX, MaxX, MinY, MaxY, MinZ, MaxZ, Seed);
     WPG.GeneratePointCloud();
@@ -273,14 +273,14 @@ TEST_WORLDPOINT_0_SEED_10_ISMATCHING(WorldPointGeneratorCuboid, Test_WorldPoint_
 ///////////////////////////////////////////////////////////////////////////////
 TEST_WORLDPOINT_4_SEED_10_ISMATCHING(WorldPointGeneratorCuboid, Test_WorldPoint_4_Seed_10_IsMatching)
 {
-    const uint64  NumberOfWorldPointsToCreate{1000U};
+    const uint32  NumberOfWorldPointsToCreate{1000U};
     const float64 MinX{-5.0};
     const float64 MaxX{5.0};
     const float64 MinY{-2.0};
     const float64 MaxY{2.0};
     const float64 MinZ{3.0};
     const float64 MaxZ{30.0};
-    const uint64  Seed{10U};
+    const uint32  Seed{10U};
 
     WorldPointGeneratorCuboid WPG(NumberOfWorldPointsToCreate, MinX, MaxX, MinY, MaxY, MinZ, MaxZ, Seed);
     WPG.GeneratePointCloud();
@@ -306,7 +306,7 @@ TEST_WORLDPOINT_4_SEED_10_ISMATCHING(WorldPointGeneratorCuboid, Test_WorldPoint_
 ///////////////////////////////////////////////////////////////////////////////
 TEST_WORLDPOINTS_INRANGE(WorldPointGeneratorCuboid, Test_WorldPoints_InRange)
 {
-    const uint64  NumberOfWorldPointsToCreate{1000U};
+    const uint32  NumberOfWorldPointsToCreate{1000U};
     const float64 MinX{-5.0};
     const float64 MaxX{5.0};
     const float64 MinY{-2.0};
@@ -319,9 +319,9 @@ TEST_WORLDPOINTS_INRANGE(WorldPointGeneratorCuboid, Test_WorldPoints_InRange)
 
     const ListColumnVectorFloat64_3d& WorldPoints{WPG.GetWorldPoints()};
 
-    uint64 NumberOfPointsOutOfRange = NumberOfWorldPointsToCreate;
+    uint32 NumberOfPointsOutOfRange = NumberOfWorldPointsToCreate;
 
-    for(uint64 i_WorldPoint{0U}; i_WorldPoint < NumberOfWorldPointsToCreate; i_WorldPoint++)
+    for(uint32 i_WorldPoint{0U}; i_WorldPoint < NumberOfWorldPointsToCreate; i_WorldPoint++)
     {
         const ColumnVectorFloat64_3d& CurrentWorldPoint{WorldPoints[i_WorldPoint]};
